requirements.txt
# Core dependencies
python-telegram-bot==20.7
openai==1.3.0

# Google APIs
gspread==5.12.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Email processing
imaplib2==3.6
email-validator==2.0.0

# Scheduling and time
apscheduler==3.10.4
pytz==2023.3
dateparser==1.1.8

# Document processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
img2pdf>=0.4.0
reportlab>=4.0.0
pdf2image==1.16.3

# Image processing and OCR (fixed numpy version for compatibility)
numpy<2.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1

# Speech processing
SpeechRecognition==3.10.0
pydub==0.25.1
edge-tts==6.1.9  # Microsoft Edge TTS for high-quality Russian voice

# RAG and ML (optimized versions)
huggingface-hub==0.16.4
sentence-transformers==2.2.2
chromadb==0.4.18

# HTTP requests for AmoCRM API
requests==2.31.0